Fixing dots; Changing how props work
I was trying to get repeats to work and noticed that they use XML
attributes, so I registered them with `sattr`, but realized that looks
for a child-node and pulls the text out of for the value.

So I changed it to `prop` but realized that's a binary value, which
confused me.

This changes `prop` to actually read the attribute value from a node.

It also changes `dot` to be a `pnode` because you can have more than one
dot, and they're defined as sub-nodes on the `note` type.